Start exiling the test-only queries into the test package.
