--- src-electron/db/query-generic.js
@@ -1,40 +1,0 @@
-/**
- *
- *    Copyright (c) 2020 Silicon Labs
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-/**
- * This module provides generic queries.
- *
- * @module DB API: generic queries against the database.
- */
-const dbApi = require('./db-api.js')
-
-/**
- * Simple query that returns number of rows in a given table.
- *
- * @export
- * @param {*} db
- * @param {*} table
- * @returns a promise that resolves into the count of the rows in the table.
- */
-async function selectCountFrom(db, table) {
-  return dbApi
-    .dbGet(db, `SELECT COUNT(*) AS COUNT FROM ${table}`)
-    .then((x) => x.COUNT)
-}
-
-// exports
-exports.selectCountFrom = selectCountFrom
