+++ test/zcl-loader.test.js
@@ -22,11 +22,11 @@ const dbApi = require('../src-electron/db/db-api.js')
 const dbEnum = require('../src-shared/db-enum.js')
 const queryZcl = require('../src-electron/db/query-zcl.js')
 const queryPackage = require('../src-electron/db/query-package.js')
-const queryGeneric = require('../src-electron/db/query-generic.js')
 const zclLoader = require('../src-electron/zcl/zcl-loader.js')
 const env = require('../src-electron/util/env.js')
 const types = require('../src-electron/util/types.js')
 const testUtil = require('./test-util.js')
+const testQuery = require('./test-query.js')
 
 test('test opening and closing the database', async () => {
   let db = await dbApi.initRamDatabase()
@@ -71,21 +71,21 @@ test('test Silabs zcl data loading in memory', () => {
     .then((x) => expect(x.length).toEqual(121))
     .then(() => queryZcl.selectAllDeviceTypes(db, packageId))
     .then((x) => expect(x.length).toEqual(175))
-    .then(() => queryGeneric.selectCountFrom(db, 'COMMAND_ARG'))
+    .then(() => testQuery.selectCountFrom(db, 'COMMAND_ARG'))
     .then((x) => expect(x).toEqual(testUtil.totalCommandArgsCount))
-    .then(() => queryGeneric.selectCountFrom(db, 'COMMAND'))
+    .then(() => testQuery.selectCountFrom(db, 'COMMAND'))
     .then((x) => expect(x).toEqual(testUtil.totalCommandCount))
-    .then(() => queryGeneric.selectCountFrom(db, 'ENUM_ITEM'))
+    .then(() => testQuery.selectCountFrom(db, 'ENUM_ITEM'))
     .then((x) => expect(x).toEqual(testUtil.totalEnumItemCount))
-    .then(() => queryGeneric.selectCountFrom(db, 'ATTRIBUTE'))
+    .then(() => testQuery.selectCountFrom(db, 'ATTRIBUTE'))
     .then((x) => expect(x).toEqual(testUtil.totalAttributeCount))
-    .then(() => queryGeneric.selectCountFrom(db, 'BITMAP_FIELD'))
+    .then(() => testQuery.selectCountFrom(db, 'BITMAP_FIELD'))
     .then((x) => expect(x).toEqual(726))
-    .then(() => queryGeneric.selectCountFrom(db, 'STRUCT_ITEM'))
+    .then(() => testQuery.selectCountFrom(db, 'STRUCT_ITEM'))
     .then((x) => expect(x).toEqual(165))
-    .then(() => queryGeneric.selectCountFrom(db, 'GLOBAL_ATTRIBUTE_DEFAULT'))
+    .then(() => testQuery.selectCountFrom(db, 'GLOBAL_ATTRIBUTE_DEFAULT'))
     .then((x) => expect(x).toEqual(126))
-    .then(() => queryGeneric.selectCountFrom(db, 'SPEC'))
+    .then(() => testQuery.selectCountFrom(db, 'SPEC'))
     .then((x) => expect(x).toEqual(testUtil.totalSpecCount))
     .then(() =>
       dbApi.dbAll(
@@ -210,11 +210,11 @@ test('test Dotdot zcl data loading in memory', async () => {
       .then((x) => expect(x.length).toEqual(41))
       .then(() => queryZcl.selectAllDeviceTypes(db, packageId))
       .then((x) => expect(x.length).toEqual(108))
-      .then(() => queryGeneric.selectCountFrom(db, 'COMMAND_ARG'))
+      .then(() => testQuery.selectCountFrom(db, 'COMMAND_ARG'))
       .then((x) => expect(x).toEqual(644)) // seems low
-      .then(() => queryGeneric.selectCountFrom(db, 'COMMAND'))
+      .then(() => testQuery.selectCountFrom(db, 'COMMAND'))
       .then((x) => expect(x).toEqual(238)) // seems low
-      .then(() => queryGeneric.selectCountFrom(db, 'ATTRIBUTE'))
+      .then(() => testQuery.selectCountFrom(db, 'ATTRIBUTE'))
       .then((x) => expect(x).toEqual(630)) // seems low
       .then(() => queryZcl.selectAllAtomics(db, packageId))
       .then((x) => expect(x.length).toEqual(56)) //This is the correct count from the ZCL8 chapter 2.6
@@ -222,11 +222,11 @@ test('test Dotdot zcl data loading in memory', async () => {
       .then((x) => expect(x.length).toEqual(61)) //seems low
       .then(() => queryZcl.selectAllEnums(db, packageId))
       .then((x) => expect(x.length).toEqual(105)) //seems low
-      .then(() => queryGeneric.selectCountFrom(db, 'ENUM_ITEM'))
+      .then(() => testQuery.selectCountFrom(db, 'ENUM_ITEM'))
       .then((x) => expect(x).toEqual(639))
       .then(() => queryZcl.selectAllStructs(db, packageId))
       .then((x) => expect(x.length).toEqual(20)) //seems low
-      .then(() => queryGeneric.selectCountFrom(db, 'STRUCT_ITEM'))
+      .then(() => testQuery.selectCountFrom(db, 'STRUCT_ITEM'))
       .then((x) => expect(x).toEqual(63))
 
       //Do some checking on the device type metadata
