+++ test/zcl-properties-loader.test.js
@@ -22,10 +22,10 @@ const dbApi = require('../src-electron/db/db-api.js')
 const dbEnum = require('../src-shared/db-enum.js')
 const queryZcl = require('../src-electron/db/query-zcl.js')
 const queryPackage = require('../src-electron/db/query-package.js')
-const queryGeneric = require('../src-electron/db/query-generic.js')
 const zclLoader = require('../src-electron/zcl/zcl-loader.js')
 const env = require('../src-electron/util/env.js')
 const path = require('path')
+const testQuery = require('./test-query.js')
 
 const zclTestPropertiesFile = path.join(
   __dirname,
@@ -64,17 +64,17 @@ test('test Silabs zcl data loading in memory', () => {
     .then((x) => expect(x.length).toEqual(120))
     .then(() => queryZcl.selectAllDeviceTypes(db, packageId))
     .then((x) => expect(x.length).toEqual(152))
-    .then(() => queryGeneric.selectCountFrom(db, 'COMMAND_ARG'))
+    .then(() => testQuery.selectCountFrom(db, 'COMMAND_ARG'))
     .then((x) => expect(x).toEqual(1663))
-    .then(() => queryGeneric.selectCountFrom(db, 'COMMAND'))
+    .then(() => testQuery.selectCountFrom(db, 'COMMAND'))
     .then((x) => expect(x).toEqual(553))
-    .then(() => queryGeneric.selectCountFrom(db, 'ENUM_ITEM'))
+    .then(() => testQuery.selectCountFrom(db, 'ENUM_ITEM'))
     .then((x) => expect(x).toEqual(1560))
-    .then(() => queryGeneric.selectCountFrom(db, 'ATTRIBUTE'))
+    .then(() => testQuery.selectCountFrom(db, 'ATTRIBUTE'))
     .then((x) => expect(x).toEqual(3408))
-    .then(() => queryGeneric.selectCountFrom(db, 'BITMAP_FIELD'))
+    .then(() => testQuery.selectCountFrom(db, 'BITMAP_FIELD'))
     .then((x) => expect(x).toEqual(721))
-    .then(() => queryGeneric.selectCountFrom(db, 'STRUCT_ITEM'))
+    .then(() => testQuery.selectCountFrom(db, 'STRUCT_ITEM'))
     .then((x) => expect(x).toEqual(154))
     .then(() =>
       dbApi.dbAll(
